+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+* Adds lookback time and distance calculations.
+* Adds comoving volume calculations.
+* Updates values for constants based on CODATA 2018.
+* Refactors handling of base units.
+
 # 0.1.0
 
 Initial release.